Fix address form for new subscription, add new VAT rules to edit billing address form
